requirements.txt
Django==5.2
djangorestframework
PyJWT
python-dotenv
pydantic
pydantic[email]
gunicorn
psycopg2-binary
whitenoise
reportlab